backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-storages==1.14.3
boto3==1.34.131
psycopg2-binary==2.9.9
gunicorn==22.0.0
Pillow==10.3.0
python-dotenv==1.0.1
